+++ config/default.go
@@ -1,28 +1,47 @@
 package config
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/style"
+	"github.com/samber/lo"
 	"github.com/spf13/viper"
 	"reflect"
 )
 
 type Field struct {
-	Name        string
+	Key         string
 	Value       any
 	Description string
+}
+
+func (f *Field) Json() string {
+	field := struct {
+		Key         string `json:"key"`
+		Value       any    `json:"value"`
+		Description string `json:"description"`
+		Type        string `json:"type"`
+	}{
+		Key:         f.Key,
+		Value:       f.Value,
+		Description: f.Description,
+		Type:        reflect.TypeOf(f.Value).String(),
+	}
+
+	output := lo.Must(json.Marshal(field))
+	return string(output)
 }
 
 func (f *Field) Pretty() string {
 	return fmt.Sprintf(
 		`%s
-%s: %s ~ %s
+%s: %s = %s
 `,
 		style.Faint(f.Description),
-		style.Magenta(f.Name),
+		style.Magenta(f.Key),
 		style.Yellow(reflect.TypeOf(f.Value).String()),
-		style.Cyan(fmt.Sprintf("%v", viper.Get(f.Name))),
+		style.Cyan(fmt.Sprintf("%v", viper.Get(f.Key))),
 	)
 }
 
@@ -33,13 +52,13 @@ func init() {
 			".",
 			`Where to download manga
 Absolute or relative.
-You can also use home variable 
-path = "~/..." or "$HOME/..."`,
+You can also use tilde (~) to refer to your home directory or use env variables.
+Examples: ~/... or $HOME/... or ${MANGA_PATH}-mangal`,
 		},
 		{
 			constant.DownloaderChapterNameTemplate,
 			"[{padded-index}] {chapter}",
-			`Name template of the downloaded chapters
+			`Key template of the downloaded chapters
 Path forbidden symbols will be replaced with "_"
 Available variables:
 {index}          - index of the chapters
@@ -209,7 +228,7 @@ Use "any" to show all languages`,
 		{
 			constant.GenAuthor,
 			"",
-			"Name to use in generated scrapers as author",
+			"Key to use in generated scrapers as author",
 		},
 		{
 			constant.LogsWrite,
@@ -285,8 +304,8 @@ panic, fatal, error, warn, info, debug, trace`,
 	}
 
 	for _, field := range fields {
-		Default[field.Name] = field
-		EnvExposed = append(EnvExposed, field.Name)
+		Default[field.Key] = field
+		EnvExposed = append(EnvExposed, field.Key)
 	}
 }