+++ constant/prefix.go
@@ -1,0 +1,6 @@
+package constant
+
+const (
+	TempPrefix  = Mangal + "Temp_"
+	CachePrefix = Mangal + "Cache_"
+)