+++ constant/template.go
@@ -1,5 +1,11 @@
 package constant
 
+const (
+	SearchMangaFn   = "SearchManga"
+	MangaChaptersFn = "MangaChapters"
+	ChapterPagesFn  = "ChapterPages"
+)
+
 const SourceTemplate = `{{ $divider := repeat "-" (plus (max (len .URL) (len .Name) (len .Author) 3) 12) }}{{ $divider }}
 -- @name    {{ .Name }} 
 -- @url     {{ .URL }}
@@ -76,4 +82,5 @@ end
 ----- HELPERS -----
 --- END HELPERS ---
 
--- ex: ts=4 sw=4 et filetype=lua`
+-- ex: ts=4 sw=4 et filetype=lua
+`