+++ provider/generic/chapters.go
@@ -1,7 +1,7 @@
 package generic
 
 import (
-	"github.com/gocolly/colly"
+	"github.com/gocolly/colly/v2"
 	"github.com/metafates/mangal/source"
 	"net/http"
 )
@@ -23,7 +23,6 @@ func (s *Scraper) ChaptersOf(manga *source.Manga) ([]*source.Chapter, error) {
 
 	if s.config.ReverseChapters {
 		// reverse chapters
-		// will happen only once for each manga
 		chapters := s.chapters[manga.URL]
 		reversed := make([]*source.Chapter, len(chapters))
 		for i, chapter := range chapters {